vectors, each loops, push/pop operations

+++ cart/includes/stack.h
@@ -3,7 +3,7 @@
 
 #include "arena_memory.h"
 
-void write_chunk(Chunk *c, uint8_t byte, int line);
+void write_chunk(Chunk *c, uint16_t byte, int line);
 int add_constant(Chunk *c, Element ar);
 
 void push(Stack **s, Element e);
